requirements.txt
# Core Dependencies
Flask==3.0.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
influxdb-client==1.38.0
python-dotenv==1.0.0
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
pydantic==2.5.0
paho-mqtt==1.6.1
PyJWT==2.8.0
bcrypt==4.1.2
redis==5.0.1
email-validator==2.1.0

# Testing Dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
pytest-mock==3.12.0
requests==2.31.0
faker==20.1.0
